refactor(signals): move release-dependent trigger to sboms/signals.py
Pre-merge cleanup from code review findings on #881:
- Moves trigger_release_dependent_assessments from core/signals.py to
  sboms/signals.py to restore correct app dependency direction (core
  should not import from plugins).
- Adds _suppress_collection_signals guard for consistency with the
  sibling bump_collection_version_on_artifact_added handler.
- Corrects return type annotation from Any to None.

Refs sbomify/sbomify#873
