+++ sbomify/apps/core/signals.py
@@ -239,89 +239,6 @@ def auto_create_component_release_on_sbom_save(sender: Any, instance: Any, creat
         logger.error("Error auto-creating ComponentRelease for SBOM %s", instance.id, exc_info=True)
 
 
-@receiver(post_save, sender="core.ReleaseArtifact")
-def trigger_release_dependent_assessments(sender: Any, instance: Any, created: Any, **kwargs: Any) -> Any:
-    """Enqueue release-dependent plugin assessments when an SBOM is linked to a named release.
-
-    Plugins marked ``requires_release=True`` (currently only dependency-track)
-    are excluded from the SBOM upload signal and instead triggered from this
-    handler. This eliminates a race condition where the SBOM upload signal
-    fires before the action's separate POST /sboms/{id}/releases call has
-    created the ReleaseArtifact row, causing dependency-track to fail with
-    "no release found" because the link doesn't exist yet.
-
-    Only fires for newly-created ReleaseArtifact rows that:
-    - link an SBOM (not a document)
-    - belong to a named release (not the auto-maintained "latest" release)
-
-    The "latest" release is skipped because every SBOM upload synchronously
-    auto-adds the SBOM to "latest" via update_latest_release_on_sbom_created,
-    which would cause DT to run twice per upload — once for "latest" and once
-    for the action's named release. Named releases are the authoritative
-    scan targets.
-
-    See spec: docs/superpowers/specs/2026-04-07-release-dependent-plugin-trigger-design.md
-    """
-    if not created:
-        return
-    if instance.sbom_id is None:
-        return
-
-    from sbomify.apps.core.models import Release
-
-    release_info = Release.objects.filter(pk=instance.release_id).values_list("is_latest", "product__team_id").first()
-    if release_info is None:
-        logger.debug(
-            "ReleaseArtifact %s has no reachable release/product/team, skipping plugin assessments",
-            instance.pk,
-        )
-        return
-
-    is_latest, team_id_value = release_info
-    if is_latest:
-        return
-
-    team_id = str(team_id_value)
-
-    from sbomify.apps.core.services.transactions import run_on_commit
-    from sbomify.apps.plugins.sdk.enums import RunReason
-    from sbomify.apps.plugins.tasks import enqueue_assessments_for_sbom
-
-    sbom_id = instance.sbom_id
-    artifact_id = instance.pk
-
-    def _enqueue() -> None:
-        try:
-            enqueued = enqueue_assessments_for_sbom(
-                sbom_id=sbom_id,
-                team_id=team_id,
-                run_reason=RunReason.ON_RELEASE_ASSOCIATION,
-                release_dependent_only=True,
-            )
-            if enqueued:
-                logger.info(
-                    "Enqueued %d release-dependent plugin assessments for SBOM %s (via ReleaseArtifact %s): %s",
-                    len(enqueued),
-                    sbom_id,
-                    artifact_id,
-                    enqueued,
-                )
-            else:
-                logger.debug(
-                    "No release-dependent plugins enqueued for SBOM %s (none enabled)",
-                    sbom_id,
-                )
-        except Exception:
-            logger.warning(
-                "Failed to enqueue release-dependent plugin assessments for "
-                "SBOM %s (message broker may be unavailable)",
-                sbom_id,
-                exc_info=True,
-            )
-
-    run_on_commit(_enqueue)
-
-
 @receiver(post_delete, sender="sboms.SBOM")
 def cleanup_component_release_on_sbom_delete(sender: Any, instance: Any, **kwargs: Any) -> Any:
     """Clean up ComponentRelease when an SBOM is deleted.
